+++ tests/testthat/test-edge-direction.R
@@ -301,7 +301,6 @@ test_that('geometry interface returns expected', {
 
 })
 
-
 test_that('NAs in coordinates return NA', {
   copyDT <- copy(DT)
 
@@ -341,3 +340,21 @@ test_that('NAs in coordinates return NA', {
   )
 })
 
+
+test_that('geometry interface can manage edges and DT both with geometry', {
+  copy_DT <- copy(clean_DT)
+  copy_edges <- copy(clean_edges)
+
+  get_geometry(copy_DT, coords = coords, crs = utm)
+
+  fusion_id(edges = copy_edges, threshold = 50)
+  cent <- centroid_fusion(edges = copy_edges, DT = copy_DT, id = id)
+
+  expect_true('direction_dyad' %in%
+                colnames(edge_direction(edges = cent, DT = copy_DT, id = id)))
+
+  # And can manage if not present
+  cent[, geometry := NULL]
+  expect_true('direction_dyad' %in%
+                colnames(edge_direction(edges = cent, DT = copy_DT, id = id)))
+})